spyglass-deployment-action/pyproject.toml
[project]
name = "spyglass-deployment-action"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0"
]
